scripts: Add ACT_PRE_RESUME stage
On Tue, Jan 10, 2017 at 03:51:10PM +0300, Pavel Emelyanov wrote:
> On 01/09/2017 05:44 PM, Cyrill Gorcunov wrote:
> > The "post-sigreturn" action is when all tasks are restored
> > but not yet restarted sitting in frozen/seized state, right
> > after final rt-sigreturn call passed.
> >
> > In vz7 this step is needed to tune up beancounters/cgroups
> > params from the script, in vanilla criu it useful as well.
>
> We have POST_RESTORE hook. If it happens too early (why?) let's
> move it into proper place?

It happens early for vz7 case: we're inside veX and continue
manipulating with resources which can lead up to problems
with beancounters and memory limits. I suspect LXC and friends
won't hit this problem for now. We could move POST_RESTORE
hook to trigger later but this will break API I think, we simply
can't be sure that someone already is using this hook in this
particular place.

What about naming it as pre-resume?