uffd.c: create lazy-pages socket in image directory
runc uses the work-dir and image-dir option to control where
criu is running. This means that criu chdir()-s to work-dir
and the lazy-pages socket cannot be found as it is in the image
directory.

I think there was a problem with too long socket names which make it
impossible to include the full path. Therefore this change tries to
fchdir() to the image directory before accessing the lazy-pages socket
and chdir()-s back to the work-dir after that.

Signed-off-by: Adrian Reber <areber@redhat.com>