pstree: Simplify get_free_pid()
The expression rb_entry(node, struct pstree_item, pid.node)
may create a fake impression, that we dereferrence pstree_item
for threads too, which is a BUG, but it's not so, because
we are only interested in its ->pid field.

But anyway, escape of pstree_item, iterate over struct pid,
which are more readable.

Signed-off-by: Kirill Tkhai <ktkhai@virtuozzo.com>