pagemap: Support async pages reading by chunks (v2)
When we collect all the read_page requests into one big preadv
call, the latter one may not read all the data in one go and
return less bytes read, than requested.

This is valid and already met in a bug :) So advance the iovec
set with the ret value and continue reading.

v2: Compilation fix on arm-s
    A little bit more debugging

https://github.com/xemul/criu/issues/271

Signed-off-by: Pavel Emelyanov <xemul@virtuozzo.com>