criu/pie/Makefile: rm old compel leftovers
Since we treat compel as an external tool, and since compel can now
provide ldflags, plugins and so on, there is no need to have this
extra code in Makefile.

Note that because of the changes, parasite/restorer is no longer rebuilt
when there are changes in compel. To me, this fits well to a vision that
compel is an external thing. This means, unfortunately, that if you have
changed compel code, you need to run 'make clean' for a proper build.

This also makes it possible to use e.g. installed compel instance.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>