crtools: resurrect --pid option for dump, v2
This option now allows us to specify a process within specified
process tree that needs to be checkpointed. For now only root
process is allowed(-t and -p options should have same PID).

This behaviour is useful for dumping master processes,
that fork it's children to do the work. I.e. one of such cases
is php process that forks workers. When one uses this option,
he agrees to possible consequences and is ready to handle
recovering from such restore on a dumpee side.

v2: fixed help line length

Signed-off-by: Ruslan Kuprieiev <rkuprieiev@cloudlinux.com>