unix: allow --external unix[ino]:stub[string]
Needles to say it's useful for debug, but it is also
very useful when you syncronise checkpoint/restore
whith a dumpee using unix socketpair and want it to
painlessly receive notification on restore.

I.e. on dump, if --external unix[ino]:foo is specified,
criu will treat specified unix socket almost the same
way it handles criu service socket. It will dump it as
closed and will put "foo" string into images, to put
that "foo" string into that socket on restore. This is
almost identical to what criu does with service socket,
but instead of putting a Criu_Resp message into it on
restore, it puts an arbitrary string.

Signed-off-by: Ruslan Kuprieiev <rkuprieiev@cloudlinux.com>