epoll: Use kcmp for targets if available
When dumping epoll targets we assume that all files
are belonging to the process which opened epoll.
In most situations it's true but not always.

The file may be received from unix socket, or say
dup()ed, added into queue and closed then. Then
such descriptor number won't be available for us
while kernel will continue using "struct file"
internally for polling.

To be able to unweave such a tie we may use
KCMP_EPOLL_TFD opcode which is designed exactly
to process epoll targets.

Because we probe every target file the dumping
procedure get deferred until all other files
present in processes tree are collected.

Note that here we only fail to dump if target
file is unaccessible or belong to another process.

For proper restore we will need to extend epoll
image to carry target sources and lift restore
code more. In next patches.

Signed-off-by: Cyrill Gorcunov <gorcunov@openvz.org>