ns: Set NS_ROOT namespaces after tasks are read
Namespaces are read in read_ns_with_hookups(),
when tasks are not read. So, root_item is NULL,
and NS_ROOT is not set for appropriate namespaces.

This patch fixes NS_ROOT after tasks are read. Also
it adds uts, ipc and cgroup namespaces for uniformity.

v2: Use macro MARK_ROOT_NS()

Signed-off-by: Kirill Tkhai <ktkhai@virtuozzo.com>