RPC: add version check interface
Instead of parsing the output of 'criu -V' this offers a RPC interface
to get CRIU's version. In a follow up patch a test script is included
to use the new interface:

 ./version.py
 Connecting to CRIU in swrk mode to check the version:
 RPC: Success
 CRIU major 2
 CRIU minor 12
 CRIU gitid v2.12-635-g6d3ae4d

This change exports the following version fields:
 * major
 * minor
 * gitid
 * sublevel (optional)
 * extra (optional)
 * name (optional)

A 'gitid' of '0' (the '0' is a string) means that it is not built from a
git checkout.

Signed-off-by: Adrian Reber <areber@redhat.com>