reg-files: Allow to skip file size checking via --skip-fsize-path option
Some files may be updated by kernel with acct() syscall so
the size kept inside image won't match one upon task exit.

Because there is no way to fetch this setting from the
kernel without its modifucation (or moreover because
we have to be able to work with older images) lets
rather allow a user to specify file path to skip size
check on.

Signed-off-by: Cyrill Gorcunov <gorcunov@openvz.org>