restore: Don't open same files to mmap mutiple times
This is symmetrical to dump -- it's typical when one file
is mapped several times in a row. On dump we don't open
the mapping's map_files links for each of such files, thus
it makes sence to do the same on restore.

This speeds restore a little bit due to lower amount of
openat() and close() calls.

Signed-off-by: Pavel Emelyanov <xemul@virtuozzo.com>