criu restore: mismatched pid debug
Sometimes we fail to restore because of PID mismatch. In this case,
we would like to know who the hell has taken our PID. Let's print
a process tree, which looks something like this (this example is
obviously from the host):

 1 /usr/lib/systemd/systemd --switched-root --system --deserialize 24
 \_482 /usr/lib/systemd/systemd-journald
 \_746 /sbin/auditd
   \_748 /sbin/audispd
     \_751 /usr/sbin/sedispatch
 \_769 /usr/lib/systemd/systemd-logind

This is debug for https://github.com/xemul/criu/issues/282.

[This is a resend of the same patch -- first time mail server broke it]

Signed-off-by: Kir Kolyshkin <kir@openvz.org>