zdtm: Add pidns02 test
1)Create a pid namespace and child reaper in it;
2)Set a specific next pid for future created process;
3)Create one more process in the namespace and kill it;
4)Wait for signal
5)Check, that NSpids of dead task remains the same.

(Note, that currently we do not sort children to have
a specific order of their creation, and child reaper
is the second in the children list, so the test fails.
If you want to try it, replace restore_before_setsid()
in create_children_and_session() on
last_level_pid(child->pid) == 1 to have a order).

Signed-off-by: Kirill Tkhai <ktkhai@virtuozzo.com>