forking: Use last_pid_mutex for synchronization during clone()
Before this patch we used flock to order task creation,
but this way is not good. It took 5 syscalls to synchronize
a creation of a single child:

1)open()
2)flock(LOCK_EX)
3)flock(LOCK_UN)
4)close() in parent
5)close() in child

The patch introduces more effective way for synchronization,
which executes 2 syscalls only. We use last_pid_mutex,
and the syscalls number sounds definitely better.

In order to synchronize with parallel CRIU, we still
take ns_last_pid flock. After talking to @xemul,
we remain the ns_last_pid file locked all the restore time,
which is done once by criu task. In further, if we need
to relax this, we may take the flock() globally on
CR_STATE_FORKING stage only, and synchronize thread
creation with flock() locally, but now we don't need that.

Signed-off-by: Kirill Tkhai <ktkhai@virtuozzo.com>