pages: Share page_ids between ns dump-helpers
After the commit 9d2e1dfebedf ("ipc: Keep shmem segments contents into
pagemap/page images"), SysV shmem pages are stored in the pages-<n>.img
files.  Where <n> is page_ids number.

The problem here is that we dump namespaces with a helper after fork().
This results in non-shared between child(s) and parent-criu counter
with the result of overwriting pages-*.img files later by helpers/criu.
This resulted in restore with corrupted shmem segments or
in the following failures on restore:
> (04.501019)      1: Error (criu/pagemap.c:265): Can't read mapping page 0: No such file or directory
> (04.637516)      1: Error (criu/pagemap.c:265): Can't read mapping page 0: No such file or directory

Allocate page_ids with shmalloc() and make it atomic.
As pre-dump, dump and page-server may run as a commands during
one service session, don't set it up early in main() - as
on the next command the value of page_ids will be preserved.
Rather do it lately during initialization of (pre-)dump/pageserver.
As (pre-)dump/pageserver rpc commands are executed under fork(),
we will not share the counter to criu.

Cc: Pavel Emelyanov <xemul@virtuozzo.com>
Signed-off-by: Dmitry Safonov <dsafonov@virtuozzo.com>